requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
banal==1.0.6
certifi==2024.2.2
charset-normalizer==3.3.2
configparser==6.0.0
contextlib3==3.10.0
deepgram-sdk==2.5.0
deepl==1.17.0
ffmpeg-python==0.2.0
filelock==3.9.0
frozenlist==1.4.1
fsspec==2023.4.0
future==1.0.0
idna==3.6
Jinja2==3.1.2
languagecodes==1.1.1
lingua-language-detector==2.0.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
pydub==0.25.1
requests==2.31.0
scipy==1.12.0
srt==3.5.3
sympy==1.12
tqdm==4.66.2
typing_extensions==4.8.0
urllib3==2.2.1
Wave==0.0.2
websockets==12.0
yarl==1.9.4
spacy==3.7.4
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
